Implement up to render for sphere
There is a big problem I've just realized. I had initially thought that using GML would produce a list of data that we could pass to the ray tracing engine. But the surface functions require the ray tracing engine to call back into the interpreter for each pixel! I worry this will be very expensive....